refactor: Add a reusable Seats component
This commit refactors the code by creating a reusable Seats component which takes in an IHall object and renders all standard and luxury seats as well as handles occupied and selected seats dynamically. The component also has an optional prop of cardSize to customize the size of the ShowtimeSeatCard components. This promotes code reuse and simplifies the SeatsPage and HallSearch components.