Refactor controllers & routes, add middleware to check order existence before deleting items
This commit refactors several controllers and routes by adding new middleware `checkToDelete` to prevent deleting a resource that is still in use by another resource. The middleware checks if the resource has any orders associated with it before being deleted. If there are any orders associated, the middleware sends back an error response without deleting the restaurant. The middleware is added to the `barController`, `cinemaController`, `genreController`, `hallController`, and `movieController`. Additionally, `updateShowtime` and `deleteTicket` functions are commented out since they are not used.